Create README.md
A dice roller game for 2 players. The game takes name of the players and then allow players to roll the dice one after another. After the roll, the player can hold the value and with each hold the total number will increase. As soon as any player hit 100, he will win!